+++ src/lib/core/ParticleHeaders.cpp
@@ -65,7 +65,7 @@ numParticles() const
 int ParticleHeaders::
 numAttributes() const
 {
-    return attributes.size();
+    return static_cast<int>(attributes.size());
 }
 
 bool ParticleHeaders::
@@ -145,10 +145,10 @@ addAttribute(const char* attribute,ParticleAttributeType type,const int count)
     ParticleAttribute attr;
     attr.name=attribute;
     attr.type=type;
-    attr.attributeIndex=attributes.size(); //  all arrays separate so we don't use this here!
+    attr.attributeIndex=static_cast<int>(attributes.size()); //  all arrays separate so we don't use this here!
     attr.count=count;
     attributes.push_back(attr);
-    nameToAttribute[attribute]=attributes.size()-1;
+    nameToAttribute[attribute]=static_cast<int>(attributes.size()-1);
     return attr;
 }
